fix: bug with writing to db
