+++ crypto_sentiment_demo_app/label_studio/requirements.txt
@@ -2,3 +2,4 @@ label-studio-sdk==0.0.12
 SQLAlchemy >= 1.4.36
 hydra-core==1.1.2
 python-dotenv >= 0.20.0
+pangres >= 4.1.1
